pyproject.toml
[tool.poetry]
name = "flask-bp"
version = "0.2.0"
description = "A comprehensive boilerplate for flask projects"
authors = ["Abdul Wakeel <awkhan978@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
flask = "=1.1.2"
flask-restful = "=0.3.8"
flask_cors = "=3.0.9"
marshmallow = "=3.7.1"
flask-sqlalchemy = "=2.4.4"
PyYAML = "=5.3.1"
flask_script = "=2.0.6"
flask_migrate = "=2.5.3"
psycopg2 = "^2.8"
flask-apispec = "0.8.6"
apispec = "3.2.0"

[tool.poetry.dev-dependencies]
pytest = "=6.2.2"
"testing.postgresql" = "^1.3"
coverage = "=5.4"
pytest-mock = "=3.3.1"
pylint = "^2.6"
pytest-cov = "=2.10.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
